+++ stag_lib/CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - [Issue #5](https://github.com/pmacg/stag/issues/5): Rename the `Graph::volume()` method to `Graph::total_volume()`
+- [Issue #5](https://github.com/pmacg/stag/issues/5): Attempting to construct a graph with an assymetric adjacency matrix
+throws an exception
 
 ### Added
 - [Issue #5](https://github.com/pmacg/stag/issues/5): `Graph::degree_matrix()` method
 - [Issue #5](https://github.com/pmacg/stag/issues/5): `Graph::normalised_laplacian()` method
+- [Issue #5](https://github.com/pmacg/stag/issues/5): `Graph::number_of_vertices()` method
+- [Issue #5](https://github.com/pmacg/stag/issues/5): `Graph::number_of_edges()` method
 
 ## [0.1.6] - 2022-10-10
 ### Added